Factor notifications into separate solids rather than hooks
